Ensure the cloud url is used when publishing a status to activitypub Posting is done in two steps - first the media is uploaded and that begins a processing pipeline. When the media is finally processed, the media pipeline sets the cdn_url to the final image (in the cloud if configured). At any time, the user can click post. This begins a separate NewStatusPipeline to deliver the status over ActivityPub. This causes a race. If the NewStatusPipeline runs before the media finishes processing, then the status is delivered with the un-processed (and un-optimized images). If pixelfed is set to use cloud storage, then the ActivityPub message also incorrectly uses the local media.
This commit fixes the race condition by waiting for all the media to be processed before sending the message over ActivityPub.
A new publish_delayed column is set per-post, indicating whether the
status should be delayed until all media is processed.

Then, the NewStatusPipeline is called twice per post -
When the media finishes processing and when the status is posted.

Both times, the pipeline will check to see if it's valid to post
If so, the pipeline will try to set publish_delayed to false
and only if it is the first time doing so, will it publish the post