Fix null dereference error when updating the sharedInbox for actor updates.
The old code assumes that all update messages have an endpoints key.
However, both the endpoints key, as well as the sharedInbox may be
missing or null per spec. This commit fixes the bug by using isset to
check for the key to exist, otherwise coalescing it to null

server to server update messages are PUT operations, not PATCH,
so if the sharedInbox is missing in the update, it should always be
removed (if previously set)